# final-project-fenwayframework/service/docs/openapi/eventify.yaml
openapi: 3.1.0
info:
  title: API for Eventify
  summary: REST APIs for Eventify
  version: "1.0"
  description: REST APIs for managing events and users
servers:
  - url: http://localhost:9000
    description: Local development environment.
paths:
...
  /user:
    get:
      operationId: get-current-user
      description: Retrieve the current authenticated user
      responses:
        200:
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        401:
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: patch-user
      description: updates the current authenticated user partially
      responses:
        200:
          description: Successfully updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        401:
          description: Authentication Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: delete-user
      description: Delete an authenticated user
      responses:
        200:
          description: Successfully deleted user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        401:
          description: Authentication Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    User:
      type: object
      required: [id, firstName, lastName, email]
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string