# final-project-fenwayframework/service/service/models/user.ts
import mongoose, { Schema, Document, Model } from "mongoose";
import bcrypt from "bcryptjs";

export interface IUser extends Document {
  firstName: string;
  lastName: string;
  location: string;
  email: string;
  password: string;
  isAdmin: boolean;
  // isAdminPassword: string;
  createdAt?: Date;
  updatedAt?: Date;
  profilePhotoURL: string;
  comparePassword(candidatePassword: string): Promise<boolean>;
}

const userSchema: Schema<IUser> = new Schema(
  {
    firstName: {
      type: String,
      required: true,
      trim: true,
    },
    lastName: {
      type: String,
      required: true,
      trim: true,
    },
    location: {
      type: String,
      required: true,
    },
    email: {
      type: String,
      required: true,
      unique: true,
      trim: true,
      lowercase: true,
    },
    password: {
      type: String,
      required: true,
      minlength: 8,
    },
    isAdmin: {
      type: Boolean,
      default: false,
    },
    profilePhotoURL: {
      type: String,
      default: ""
    }
    // isAdminPassword: {
    //   type: String,
    //   default: "",
    //   required: function() {
    //     return this.isAdmin === true;
    //   }
    // }
  },
  { timestamps: true }
);

// Pre-save middleware to hash the password if modified
userSchema.pre("save", async function (next) {
  if (this.isModified("password")) {
    this.password = await bcrypt.hash(this.password, 10);
  }
  next();
});

// Method to compare passwords
userSchema.methods.comparePassword = async function (
  candidatePassword: string
): Promise<boolean> {
  return bcrypt.compare(candidatePassword, this.password);
};

// Create and export the User model
const User: Model<IUser> = mongoose.model<IUser>("User", userSchema);
export default User;